+++ livecomponents/__init__.py
@@ -1,4 +1,14 @@
 from livecomponents.component import LiveComponent, command
-from livecomponents.manager.manager import CallContext, InitStateContext
+from livecomponents.manager.manager import (
+    CallContext,
+    InitStateContext,
+    UpdateStateContext,
+)
 
-__all__ = ["LiveComponent", "command", "CallContext", "InitStateContext"]
+__all__ = [
+    "LiveComponent",
+    "command",
+    "CallContext",
+    "InitStateContext",
+    "UpdateStateContext",
+]