+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## UNRELEASED
+
+- Made UpdateStateContext importable directly from livecomponents
+
 ## 1.4.0 (2023-11-10)
 
 - Made it possible to re-render the component state on re-render.